feat: apply global modal provider
