+++ src/app/layout.tsx
@@ -1,6 +1,7 @@
 import { GlobalFooter, GlobalHeader, GlobalMain } from 'src/components/blocks';
 import AppProvider from 'src/context/AppProvider';
 import { AuthProvider } from 'src/context/AuthProvider';
+import ModalProvider from 'src/context/ModalProvider';
 
 import '../constants/styles/global.css';
 
@@ -14,11 +15,13 @@ export default function RootLayout({
       <body>
         <AuthProvider>
           <AppProvider>
-            <div className="flex flex-col items-center min-h-screenflex">
-              <GlobalHeader />
-              <GlobalMain>{children}</GlobalMain>
-              <GlobalFooter />
-            </div>
+            <ModalProvider>
+              <div className="flex flex-col items-center min-h-screenflex">
+                <GlobalHeader />
+                <GlobalMain>{children}</GlobalMain>
+                <GlobalFooter />
+              </div>
+            </ModalProvider>
           </AppProvider>
         </AuthProvider>
       </body>
