+++ src/app/notice/page.tsx
@@ -1,15 +1,12 @@
 import React from 'react';
 
 import { NoticeBanner, NoticeSearchArea } from 'src/components/pages/Notice';
-import NoticePageProvider from 'src/context/NoticePageProvider';
 
 const NoticePage: React.FC = () => {
   return (
     <div className="w-full">
       <NoticeBanner />
-      <NoticePageProvider>
-        <NoticeSearchArea />
-      </NoticePageProvider>
+      <NoticeSearchArea />
     </div>
   );
 };
